Create the module of WordAttribute. Clean the project structure.
